Merged PR 1380: [1/n] Prefill Chunking - abstractions and class hierarchies
**PR series for porting Sarathi on top of latest vLLM**

This is a first in a series of PRs aimed at adding support for prefill chunking in vLLM. Prefill chunking and decode maximal batching are the two techniques serving as the cornerstone for Sarathi.

In this PR, we've modified the various request and request metadata abstractions to incorporate a notion of prefill chunk sizes. Scheduler and block manager class hierarchies are also implemented. Additional changes include wiring up changes to LLMEngine.