Fix docker python version
Docker images were broken because we build the extensions with
python3.8 but then we try to run them with other version of python.

This results in exception error

  File "/workspace/vllm/model_executor/layers/quantization/awq.py", line 6, in <module>
    from vllm._C import ops
ModuleNotFoundError: No module named 'vllm._C'

I think if we want to make sure vllm works in older versions of python
we should do something else.

Also fix for the cupy dependency
The docker build was not working with the requirement file