Add code-revision argument for Hugging Face Hub
Introduce the `code-revision` argument to facilitate the selection of a specific revision when fetching model code from the Hugging Face Hub. Prior to this enhancement, the default behavior was to always download the latest version which may not always be intended. Users can now specify a revision to ensure consistent model code retrieval. The default behavior remains unchanged, automatically fetching the latest version when the argument is not provided.

Run yapf and ruff