Make ray optional for single-node deployment
ray is a powerful platform for general purpose distributed computing but potentially overkill for the specific requirements of realtime synchronized inferencing between GPUs on a single node.

We would prefer to have a "lightweight" option without the ray dependency for non-ray cluster environments.

With the changes in this PR, ray will continue to be used for parallel workers if it's installed, otherwise vanilla python multiprocessing is used.

Worker processes are shut down when the LLMEngine is garbage collected.

Co-authored-by: Sahil Suneja <suneja@us.ibm.com>