Benchmark : Turn-off nightly multi-gpu benchmarks temporarily (#130)
SUMMARY:
The nightly benchmarks run on `A10g x 4` and `A10g x 1` instances. The
`A10gx4` instance takes about 6.5 hrs and the `A10g` instance is
estimated to take about 8.5 hrs.

This PR turns-off the `A10g x 4` instance for the following reasons,
 * cut costs
* The models we benchmark with don't use a `A10g x 4` in a realistic
scenario.

The `A10g x 4` instance shouldn't take `6.5` hrs to run the nightly. It
is likely there are optimization that we can make. This is a temporary
solution until we investigate.

TEST PLAN:
No testing.

Co-authored-by: Varun Sundar Rabindranath <varun@neuralmagic.com>