Benchmark Fix : Fix JSON decode error (#142)
Summary:

Bug : In benchmark_serving.py we use aiohttp's `iter_any()` function for
reading responses from the server.
Sometimes, the data received from this function contains more than one
JSON in it. This leads to a JSON decode error.

Fix : These JSONs essentially contain the generated text and the last of
the JSONs has the entire generated text. The
fix is to identify the last JSON in the response data. The response data
simply puts the JSONs next to each other. We identify occurrences of the
string `{"text":` in the `body` and use the JSON decoder to identify the
last JSON.

Test:
Manual testing

---------

Co-authored-by: Varun Sundar Rabindranath <varun@neuralmagic.com>