Benchmark : Fix remote push job (#129)
SUMMARY:
`github-action-benchmark` action, needs a JSON file with metrics for
reporting. It throws an error when the JSON is empty or doesn't have any
data.

Bug: On the `remote push` benchmark job, we produce the necessary files,
but they don't have any JSON data.

Fix: Make the logging script skip file creation if the file is going to
be empty. In the GHA side, add logic to skip processing when a desired
file does not exist.

Additional changes:
 - Rename `GHABenchmarkToolName` -> `BenchmarkMetricType`
- Add a `Observation` BenchmarkMetricType - This could be useful in the
near future when we discover volatile metrics.

TEST PLAN:
Jobs on this PR.

Co-authored-by: Varun Sundar Rabindranath <varun@neuralmagic.com>