whl centric workflow for "remote push" (#117)
SUMMARY:
* moves "nm-vllm" to whl centric workflows. this means that testing is
conducted on a fresh instance via installing "nm-vllm" whl into a venv.
* update build action to also generate whl and source distributions.
* update testing action to work via whl install.
* updates callable build workflow to build, generate whl, and source
distribution. artifacts are uploaded to GHA artifact store.
* updates callable test workflow to install whl and run tests.
* patch code coverage to work with "forked" processes

NOTES: currently we are building on a GPU instance, but technically this
isn't a requirement. the current non-gpu instance we have in our runner
group doesn't have enough memory. this results on builds sporadically
failing, so "remote push" is set up to run on T4's.

TEST PLAN:
runs on remote push

---------

Co-authored-by: andy-neuma <andy@neuralmagic.com>