[Bugfix] Fix marlin 2:4 kernel crash on H100 (#243)
The reason for the crash was the inline PTX assembly that introduced the
async_copy with streaming behavior. The solution is to use the more
standard PTX for async_copy (without the fractional L2 policy for
"evict_first"). There is no performance difference between standard
async_copy PTX and the previous one.
Ported from dense marlin:
https://github.com/vllm-project/vllm/pull/4218/