[BugFix] Don't start a Ray cluster when not using Ray
The logic to auto-select the distributed executor backend uses ray.util.get_current_placement_group() to check whether it's already running in a Ray placement group, but this function has a side effect of starting a Ray cluster if it's not already started.

The result is that a Ray cluster gets started even when defaulting to the multiprocessing distributed backend.