Fix Phi-3 Long RoPE scaling implementation
When evaluating our fine tunes of Phi-3, we noticed a big difference in the logits outputted by Huggingface transformers and vLLM.

The most significant cause of this was a deviation in the positional embedding implementation of vLLM. Since Microsoft contributed their Phi-3 implementation to Huggingface transformers, I believe that is the reference implementation to follow:-

https://github.com/huggingface/transformers/blob/9ba9369a2557e53a01378199a9839ec6e82d8bc7/src/transformers/models/phi3/modeling_phi3.py#L153

The above link shows how Phi-3's rotation matrix is scaled. I've added similar code to vLLM's `Phi3LongRoPEScaledRotaryEmbedding` and removed the hardcoded scaling factors.