[PoC][Spec Decoding] Stop speculating when confidence is low
Just an experiment, this gives an additional ~3% speedup in a greedy bs=1 example.

This just "pads" out proposed tokens up to the num_predict_tokens with the final predicted token. Would be good if we could change the framework to allow the proposer model to return a variable number of tokens.

I'm not sure how this will work in non-greedy case - we'd still probably want to evaluate the cutoff based on a greedily sampled token.