[Speculative Decoding] MLPSpeculator Tensor Parallel support (1/2)
This PR adds support for a draft worker with TP==1 and a target worker
with TP>1. Support for draft worker>1 will come in a 2nd PR.

This PR makes use of #5414 to wrap the `MLPSpeculatorWorker` with a
`SmallerTPProposerWorker`.

Adds a test case for
`ibm-granite/granite-3b-code-instruct{-accelerator}` to `test_draft_model_tp_lt_target_model_tp2`.