[BugFix] Avoid unnecessary Ray import warnings
Currently the logs are polluted with multiple Ray import warnings even in cases where Ray is not being used. There should be an error only in the case that Ray is configured and/or required but not available, and in that case it should be fatal.

In our own builds/tests this was also interfering with the subprocess-based custom all-reduce p2p check recently introduced in https://github.com/vllm-project/vllm/pull/5669.