[Bugfix] Use templated datasource in grafana.json to allow automatic imports
This is somewhat of a followup to #4711 which introducted a input variable to
allow setting the datasource on import. Unfortunately this only works when
the dashboard is imported via the Grafana UI.
It's common to e.g. import dashboard JSON files via ConfigMaps on Kubernetes,
so there is no manual step to set the datasource name.

Switching to a template var for the datasource also should cover the issue #4711
fixed, but also works in setting the datasource when using automatic dashboard
provisioning.

Signed-off-by: Christian Rohmann <christian.rohmann@inovex.de>