Fix the lm_head in gptbigcode in lora mode
Due to the weight-tie, in lora mode the lm_head implementation
was being replaced with VocabParallelEmbeddingWithLoRA which is not
meant for the lm_head. To fix the issue this commit initializes
the lm_head as an instance of ParallelLMHead and assigns it the
weights of the embedding module. Due to previous differences in the
padding of the two layers (64 vs 256), the same padding is now applied
to both. The resulting vocabulary dimension had to be added to
bgmv_config.h

Signed-off-by: Max de Bayser <mbayser@br.ibm.com>