[Core] Add engine option to return only deltas or final output
The LLMEngine and AsyncLLMEngine APIs will currently return/stream cumulative outputs for all sequences at every step.

This is more data than needed for LLM.generate or the OpenAI server APIs:
- For LLM.generate and non-streaming APIs we only need the final output
- For streaming APIs we only require deltas

This PR adds an `output_kind` parameter to SamplingParams with an enum value of either CUMULATIVE, DELTA, or FINAL_ONLY.

It will reduce the number of objects that need to be constructed at each step, and the amount of data to be serialized to return to the newly-decoupled front-end API process.