[BugFix] Raise all exception variations in async generator
The async generator returned by AsyncLLMEngine generate() uses a queue of either request outputs and/or exceptions.

But the logic consuming requests from the queue only raises instances of Exceptions, not Exception classes (which are also raisable in python). This means that the CancelledError class instance we add to the queue elsewhere currently isn't being raised from the generator as it should be.