[Spec Decoding] Use target model max length as default for draft model
We currently assume a hard-coded default of 2k for the max model length in the case that it can't be inferred from the model's configuration.

In the case of spec decoding draft models, it would instead be better to assume the same max length of the target model in this case.

This is particularly helpful for MLPSpeculator models whose length can't be determined from their config.json files and should be assumed to match that of the corresponding target model that they were specifically trained for.