[Core] llm support external swapper configuration
When we create the llm engine, we support
specifying the type of external swapper
and the size of the media type. Currently,
only local files are supported. Developers
can customize the format type of the external
swapper. For example, the format of the file
is "file://path/to/directory".
And supports users to add their own external swapper.

Change-Id: I7989e7aba32ad218629c067fc984c8744c25ab64
Signed-off-by: Changqi Lu <luchangqi.123@bytedance.com>