[Core] Executor support external swapper.
Firstly, two methods determine_num_external_available_blocks and
initialize_external_cache are added to determine the number of
external blocks and perform initialization operations when
vllm-engine is started.

Then, according to the scheduling results of shceduler, the blocks
that need to be swapped are divided into two types: CPU and External,
and then the block swap operation is performed.

Change-Id: I155439e6a8af21ae2241c3eba892f82b7c03fcb2
Signed-off-by: Changqi Lu <luchangqi.123@bytedance.com>