[Core] Schedule support external swapper
In the previous version of schedule, the blocks_to_swap_out field
in the return result only has the mapping of GPU to CPU block id.
However, after adding devices other than CPU, it is necessary to
distinguish different devices, so device related information is
added to this field.

At the same time, external swapper related methods are added to
scheduler and BlockManager, and corresponding implementations are made.

Change-Id: Ia508a7092640900cdac7fe9925a1141ed851f337
Signed-off-by: Changqi Lu <luchangqi.123@bytedance.com>