[Misc] Don't dump contents of kvcache tensors on errors
https://github.com/vllm-project/vllm/pull/8305 was recently added to dump model running inputs when encountering a fatal error.

If this happens during decode however it will include the kvcache tensors which are typically huge (~60GB in the case I was testing), and can therefore take minutes to write to disk.

When this happens the engine loop is blocked and health checks time-out causing the server to be killed.

This change replaces kvcache tensors with their dtype + shape. With this the pickling is sub-second and the filesize in my test case was 7KB.