[Core] zmq: bind only to 127.0.0.1 for local-only usage
When running with `--tensor-parallel-size` greater than `1`, I noticed
that the zmq socket opened for broadcasting to workers was listening on
all interfaces. This change makes both the listening-side and (XPUB) and
the receiving side (SUB) use `127.0.0.1` instead of `*` when we know the
usage is local-only.

I also added debug logging to show which port numbers are in use by zmq.

By listening on all interfaces, an external party could connect and
subscribe to data from the `XPUB` socket if firewall settings allow it.
The data received is not particularly useful (it's tensor metadata), but
it's still unexpected and undesirable behavior to allow connections from
a wider set of potential sources than necessary.

It is already possible to control which IP address is used for the
remote case by setting the `VLLM_HOST_IP` environment variable.

Signed-off-by: Russell Bryant <rbryant@redhat.com>