[Core] Use `spawn` when cuda is already initialized
One condition that we know will be broken with the default
multiprocessing method of `fork` is if a user of vllm as a library
initializes cuda prior to running vllm. This change detects this case,
emits a warning to the log, and force sets the method to `spawn`.

Similar code exists elsewhere (for AMD, Intel) to force the use of
`spawn` in all cases for those accelerators.

We retain the default behavior if the env var is not set and cuda is
not initialized. This seems to work fine and avoids potentially
breaking code using vllm as a library without protecting their code
under `if __name__ == "__main__"`.

Signed-off-by: Russell Bryant <rbryant@redhat.com>