[Core] Avoid metrics log noise when idle
Prior to this change, there was always a log message every 10 seconds,
even when the system is idle. The noise when the metrics are all zeros
doesn't seem very interesting, so this change makes it so the log
message is only emitted while active. It will log a message whenever
metrics are non-zero, or when returning to idle from non-idle.
Repeated messages during idle now go to DEBUG instead of INFO.

Signed-off-by: Russell Bryant <rbryant@redhat.com>