Add stream support for Granite 20b Tool Use
This commits builds on previous work by Will Eaton
and adds support for streaming. It also adds the
model to the tool use unit tests.

In this commit the tool parser is renamed from
simply granite to granite-20b-fc to differentiate
from other granite models.

Another minor change is that in the chat template
the function description using the function signature
is now optional.

Co-authored-by: Will Eaton <me@wseaton.com>

Signed-off-by: Max de Bayser <mbayser@br.ibm.com>