[CI/Build]: mypy: drop duplicate coverage
I had all directories listed in `tools/mypy.sh` to ensure they worked
with `--follow-imports skip`. This change includes the last tweaks
necessary for all of those to pass.

As a next step, remove the ones that are already running
with `--follow-imports silent` based on the config in
`pyproject.toml`. Following commits will start resolving errors that
occur when adding more coverage here.

Signed-off-by: Russell Bryant <rbryant@redhat.com>