[CI/Build] Use commit hash references for github actions
This changes the references to github actions to use commit hashes
instead of tag names. These will be updated automatically by
dependabot.

GitHub recommends this approach as a security best-practice. The
reason is that a bad actor can modify what a tag points to if access
to a repository is compromised. The hash is a more concrete reference.

For more info on this, see "Referencing actions" in this blog post:
https://securitylab.github.com/resources/github-actions-building-blocks/

Signed-off-by: Russell Bryant <rbryant@redhat.com>