[BugFix] Typing fixes to RequestOutput.prompt and beam search
This reverts a recent typing change that was made in https://github.com/vllm-project/vllm/pull/9267 which was due to another recent incorrect change in the "external" beam search implementation. It would be a breaking change to the type of RequestOutput.prompt in the API (and caused some typing breakages to the front-end API code which references this field).

I have made some related changes/fixes to the beam search impl including properly returning logprobs, but it still needs quite a bit more work overall (for subsequent PRs).