Remove if blocks smaller than bs in generate_decode_buckets (#412)
With this check while running decode_block_bucket_min=128 and bs>128 it
will skip buckets smaller than bs. Then during the run buckets that got
skipped can be used by vllm and are being warmed-up which is causing
perf drop & they are not run as hpu graphs.

This change is removing said check.