[Doc] Move CONTRIBUTING to docs site
The `CONTRIBUTING.md` file has some good information for contributors.
It would be nice to have a place to organize additional documentation
aimed at contributors, so move this content over to the docs site to
seed a new `docs/source/contributing` directory.

Replace `CONTRIBUTING.md` with a link to the new home for this
content.

Signed-off-by: Russell Bryant <rbryant@redhat.com>