[Doc] Propose a vulnerability management team
The project has a policy for how vulnerabilties are reported, but
there is no specific indivudual(s) who has the responsibility for
ensuring that these reports are acted on in a timely manner. To
address this, I propose naming a "vulnerability management team" who
would have this responsibility.

The list of individuals that would seed this team is TBD.

Signed-off-by: Russell Bryant <rbryant@redhat.com>