[CI/Build] Improve mypy + python version matrix
The `mypy` CI job ran using multiple Python versions, though it was
always checking for compatibility with Python 3.9 because of the
config option in `pyproject.toml`.

This change makes it so local usage of `format.sh` will still default
to checking for compatibility with the lowest Python version (3.9).
The CI job will now check for compatibility with each python version
used in the matrix instead of only running `mypy` itself with
different versions, but always checking 3.9 compatibility.

Signed-off-by: Russell Bryant <rbryant@redhat.com>