[Frontend] Adjust try/except blocks in API impl
These were changed to separate blocks in https://github.com/vllm-project/vllm/pull/9759 but I feel it's cleaner/clearer as a single block. It actually doesn't matter which parts of the block raise the specific exceptions in the except clauses, we still want to handle them in the same way.

Signed-off-by: Nick Hill <nhill@redhat.com>