[CI/Build] change conflict PR comment from mergify
On PRs opened by dependabot, this comment is interpreted by dependabot
as a command to rebase the PR. We don't need the bots telling each
other what to do. An example is #9746.

The simplest change is to update the text so it will no longer be
interpreted as a command by dependabot. We do that by putting the
"rebase" word prior to the username mention.

Signed-off-by: Russell Bryant <rbryant@redhat.com>