[CI/Build] Always run mypy
If you would like to configure the repo to block auto-merges if the
mypy workflow fails, you will first need to merge this PR so that the
workflow runs on all PRs, regardless of the files that have changed.
This is because github configuration does not support blocking on
workflows that run conditionally.

Signed-off-by: Russell Bryant <rbryant@redhat.com>