[Doc] Update help text for `--distributed-executor-backend`
The help text for `--distributed-executor-backend` indicated that the
option would deafult to `ray` any time multiple GPUs are used. This
did not reflect the current logic for choosing a default. When
multiple GPUs can be used on a single host, it will default to `mp`
instead of `ray`, except in the case of `hpu` and `tpu`.

Signed-off-by: Russell Bryant <rbryant@redhat.com>