[Doc] Add the start of an arch overview page
My intent with this page is to provide a high-level architectural
overview of the major components of vLLM by taking a look one layer at
a time. So far, I have the first two sections: entrypoints and the LLM
engine.

The diagrams here were build with [excalidraw](https://excalidraw.com),
which embeds diagram metadata into the PNG file, allowing others to
load the image back into the tool and edit the diagram in the future
if needed.

Signed-off-by: Russell Bryant <rbryant@redhat.com>