[HPU] Add mark_step configurable for the decoder layer.  (#525)
We are seeing 10% performance regression in the llama-based model due to
https://github.com/vllm-project/vllm/pull/10239. The mark_step()
function needs to be configured differently for each model to achieve
the best performance. For some models, mark_step() for every decoder
step would be optimal, but for other models, it's better to run it every
n-th step. We are adding a counter to only register the hook for every
n-th step, which can be configured with VLLM_CONFIG_HIDDEN_LAYERS