[Doc] Add github links for source code references
Previously, the use of `sphinx.ext.viewcode` generated new pages that
included the raw source code. This change adds the
`sphinx.ext.linkcode` extension, which allows defining a custom method
for generating source code URLs given a code reference. This now
includes links to the appropriate file and line number on github for
code references.

For now I left the `viewcode` extension in place. The result is that
code references now have 2 `[source]` links - one for the copy of the
source in the docs, and one to github. If the github links seem to
work well enough, we can drop `viewcode` later.

Signed-off-by: Russell Bryant <rbryant@redhat.com>